requirements.txt
langchain==0.3.26
langchain-pinecone==0.2.8
langchain-groq==0.3.2
langchain-community==0.3.26
langchain-huggingface==0.1.2
langchain-core==0.3.55
pinecone==6.0.2
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
streamlit==1.43.2